+++ Sources/Engine/Network/CommunicationInterface.cpp
@@ -291,7 +291,7 @@ void CCommunicationInterface::PrepareForUse(BOOL bUseNetwork, BOOL bClient)
       cm_strName = net_strLocalHost;
       cm_ulLocalHost = StringToAddress(cm_strName);
       // if invalid
-      if (cm_ulLocalHost==0 || cm_ulLocalHost==-1) {
+      if (cm_ulLocalHost == 0 || cm_ulLocalHost == (ULONG)-1) {
         cm_ulLocalHost=0;
         // report it
         CPrintF(TRANS("  requested local address is invalid\n"));
